Pretend valid video config exists at startup
According to the documentation, the BIOS will
always initialize a valid video configuration
before the application is started, so applications
can reasonably expect to be able to inherit that
mode by just querying it or making small
modifications to it and leaving other parameters
untouched. Some of the SDK sample programs, such
as the MGL samples, rely on this behavior.

To accomodate this, if no video mode has been
configured when _VidConfig() is called, initialize
and enable video in Nuance's default state if the
application doesn't specify a main display config.
Separate booleans still track whether the actual
main plane or overlay plane contain valid surfaces.