Restore ability to switch "disassembly MPE"
The control panel dialog will disassemble and
display the next instruction packet for the
currently selected "disassembly MPE" which is
initialized to MPE 3. There was code to switch
this to any MPE when that MPE's text label on the
side of the dialog was clicked, but it was broken
because the corresponding static text controls did
not have the "Notify" style set on them in the
corresponding resource file. Fixing up the style
allows this to work again.

As a bonus, this change also modifies the prior
improved breakpoint handling change to
automatically switch the disassembly MPE to
whichever MPE triggered the breakpoint when a
breakpoint is hit.