Rework support for logging kprintf to a file
Introduce the new config file option "DebugLogFile"
that specifies a log file name, replacing the
kprintf option that took an integer to select
message box or file loggin. If set, Nuon will
open the specified file and append strings passed
to kprintf() to it. The messagebox-based logging
is removed, to be replaced with logging to the
status dialog in a subsequent change.