Add .editorconfig, update .gitignore
Add a .editorconfig file that configures the
VisualStudio source editor to use the prevailing
indentation style (spaces, 2 spaces for
indentation).

Update .gitignore to ignore Nuance.aps, which is
apparently a binary form of Nuance.rc used to
improve "designer efficiency" somehow, and is
recommended to be ignored by source control tools.