Finish kprintf-to-log-file support
Round out what is probably the world's laziest
implementation of printf-style format conversion
by parsing most of the available format conversion
tokens defined in the glibc printf man page on my
system. Things that are omitted:

-Long double support. Seems unlikely to be useful
 in Nuon programs.

-The 'm$' arbitrary parameter index specifier
 syntax, which looks to be a Unix-only extension
 of the C syntax.

-The 'n' format specifier, which writes the
 current number of characters written to the next
 parameter. I've never encountered a use for this,
 and it would take more work to implement a
 special case for it.

-Anything listed as deprecated or "Do not use" in
 the man page.

-ptrdiff_t (%t) type conversion. Could easily be
 added, but callers can also trivially use %z
 instead.

-errno (%m) type conversion. Difficult to get
 right on a host-side implementation without
 acccess to the target's libc implementation.
 Just do it manually in the caller if you need it.

Pretty much everything else should work
equivalently to how it works in MSVC's sprintf
implementation, which this code utilizes for the
actual format conversions. However, the error
handling hasn't been thoroughly tested and likely
isn't that robust, so it's possible you could
confuse it by passing in invalid format conversion
strings. It's meant to be good enough for a
debugging tool, which kprintf is, not for
production use, so don't copy it somewhere else
and expect great things.