Implement kprintf display in status dialog
In addition to the file logging of kprintf data,
add a new button to the status dialog that
displays the last 32 lines of kprintf output when
clicked. The data is continuously updated when
the display is active.

Document the usage in the sample nuance.cfg and
in Readme.txt. See the section "printf Debugging
(kprintf) Support."

Fixes issue #20